+++ packages/contracts/contracts/BalanceManager.sol
@@ -13,13 +13,13 @@ import {Utils} from "./libs/Utils.sol";
 import {AssetUtils} from "./libs/AssetUtils.sol";
 import {OperationUtils} from "./libs/OperationUtils.sol";
 import "./libs/Types.sol";
-import "./OperationReentrancyGuard.sol";
+import "./NocturneReentrancyGuard.sol";
 
 contract BalanceManager is
     IERC721ReceiverUpgradeable,
     IERC1155ReceiverUpgradeable,
     CommitmentTreeManager,
-    OperationReentrancyGuard
+    NocturneReentrancyGuard
 {
     using OperationLib for Operation;
 
@@ -39,15 +39,22 @@ contract BalanceManager is
         address wallet,
         address subtreeUpdateVerifier
     ) internal onlyInitializing {
-        __OperationReentrancyGuard_init();
+        __NocturneReentrancyGuard_init();
         __CommitmentTreeManager_init(subtreeUpdateVerifier);
         _wallet = IWallet(wallet);
+    }
+
+    modifier notErc721(EncodedAsset calldata encodedAsset) {
+        (AssetType assetType, address assetAddr, uint256 id) = AssetUtils
+            .decodeAsset(encodedAsset);
+        require(assetType != AssetType.ERC721, "not erc721");
+        _;
     }
 
     function _addToAssetPrefill(
         EncodedAsset calldata encodedAsset,
         uint256 value
-    ) internal {
+    ) internal notErc721(encodedAsset) {
         bytes32 assetHash = AssetUtils.hashEncodedAsset(encodedAsset);
         _prefilledAssetBalances[assetHash] += value;
 
@@ -195,7 +202,7 @@ contract BalanceManager is
         bytes calldata // data
     ) external override returns (bytes4) {
         // Must reject the transfer outside of an operation processing
-        if (reentrancyGuardStage() == NO_OPERATION_ENTERED) {
+        if (reentrancyGuardStage() == NOT_ENTERED) {
             return 0;
         }
 
@@ -218,7 +225,7 @@ contract BalanceManager is
         bytes calldata // data
     ) external override returns (bytes4) {
         // Must reject the transfer outside of an operation processing
-        if (reentrancyGuardStage() == NO_OPERATION_ENTERED) {
+        if (reentrancyGuardStage() == NOT_ENTERED) {
             return 0;
         }
 
@@ -241,7 +248,7 @@ contract BalanceManager is
         bytes calldata // data
     ) external override returns (bytes4) {
         // Must reject the transfer outside of an operation processing
-        if (reentrancyGuardStage() == NO_OPERATION_ENTERED) {
+        if (reentrancyGuardStage() == NOT_ENTERED) {
             return 0;
         }