+++ packages/contracts/contracts/test/BalanceManager.t.sol
@@ -40,7 +40,7 @@ contract BalanceManagerTest is Test {
 
     // Check storage layout file
     uint256 constant OPERATION_STAGE_STORAGE_SLOT = 74;
-    uint256 constant ENTERED_EXECUTE_ACTIONS = 3;
+    uint256 constant ENTERED_EXECUTE_ACTIONS = 4;
 
     TestBalanceManager balanceManager;
     Wallet wallet;